requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
Django==3.2
django-adminlte3==0.1.6
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
h11==0.13.0
idna==3.3
outcome==1.2.0
psycopg2-binary==2.9.3
PySocks==1.7.1
python-dotenv==0.20.0
pytz==2022.2.1
PyVirtualDisplay==3.0
requests==2.28.1
selenium==4.4.3
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.3
wsproto==1.2.0
